User can see 'liked' dogs.
When a user clicks on the Likes tab, we make a GET request against '/connections'. The server returns the results all the pets in the database that the user is connected to.

At the moment, the server is returning fake data, but the client-side script should be able to handle the real data as well.